pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "signet"
version = "0.1.0"
description = "Persistent AI research agent with KAIROS-style memory and personality"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "Pete", email = "pete@q33north.com" },
]
dependencies = [
    "anthropic>=0.50",
    "discord.py>=2.4",
    "pydantic>=2.10",
    "pydantic-settings>=2.0",
    "typer>=0.12",
    "rich>=13.0",
    "httpx>=0.27",
    "pyyaml>=6.0",
    "structlog>=24.0",
]

[project.optional-dependencies]
db = [
    "asyncpg>=0.30",
    "pgvector>=0.3",
    "sentence-transformers>=3.0",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.8",
]

[project.scripts]
signet = "signet.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/signet"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
line-length = 100
